Add files via upload
This is what I have for the UML, I didn't include the getFirst method since we probably won't have that in the final version. I'm realizing now that I didn't put in the constructor (which I think she wants) so I will put that in. Let me know if there are any mistakes or things you would change