Add files via upload
I added the Tester to the UML, added the constructor, and changed a bunch of the methods to private. I indicated that the Tester instantiates Onion. I changed the equals method and utilized the toString method and .equals to compare the binary numbers. Then, I thought it might be a good idea in the tester to just have some quick code testing out the equals method, as sometimes I get marked off for not using/testing all of the methods I've written. You can change it as you see fit, there might be a simpler way to test it. I also changed the spelling of onionRecusive ----> onionRecursive